Split up the UnpluggedStatic class
