+++ PluggableUnplugged/UnpluggedStatic.php
@@ -4,10 +4,10 @@ declare(strict_types=1);
 /**
  * Replacements for some (not all) of the WordPress pluggable.php functions.
  *
- * @package AWonderPHP/PluggableUnplugged
+ * @package AWonderPHP
  * @author  Alice Wonder <paypal@domblogger.net>
  * @license https://opensource.org/licenses/MIT MIT
- * @version 0.31
+ * @version 0.32
  * @link    https://github.com/AliceWonderMiscreations/AWonderPHP
  */
 
@@ -18,169 +18,13 @@ namespace AWonderPHP\PluggableUnplugged;
  */
 class UnpluggedStatic
 {
-    /**
-     * Takes a valid domain name and returns punycode variant, assuming the
-     * idn_to_ascii function is available.
-     *
-     * GIGO function, invalid domain name will not throw exception.
-     *
-     * @param string $domain The domain name to translate into punycode.
-     *
-     * @return string The ascii punycode version of the domain name.
-     */
-    public static function punycodeDomain(string $domain): string
-    {
-        if (function_exists('idn_to_ascii')) {
-            $domain=idn_to_ascii($domain, 0, INTL_IDNA_VARIANT_UTS46);
-        }
-        return $domain;
-    }//end punycodeDomain()
-
-    /**
-     * Takes a valid ascii domain name and returns UTF-8 variant, assuming the
-     * idn_to_utf8 function is available.
-     *
-     * GIGO function, invalid domain name will not throw exception.
-     *
-     * @param string $domain The domain name to translate into utf8.
-     *
-     * @return string The utf8 variant of the domain name.
-     */
-    public static function unpunycodeDomain(string $domain): string
-    {
-        if (function_exists('idn_to_utf8')) {
-            $domain=idn_to_utf8($domain, 0, INTL_IDNA_VARIANT_UTS46);
-        }
-        return $domain;
-    }//end unpunycodeDomain()
-
-    /**
-     * Takes a valid international e-mail address and return user@ punycode variant,
-     * assuming idn_to_ascii is available.
-     *
-     * GIGO function, invalid e-mail will not throw exception.
-     *
-     * @param string $email The e-maill address with domain name to translate into utf8.
-     *
-     * @return string The e-mail with ascii variant of the domain name
-     */
-    public static function punycodeEmail(string $email): string
-    {
-        if (substr_count($email, '@') === 1) {
-            $tmp=explode('@', $email);
-            $email=$tmp[0] . '@' . self::punycodeDomain($tmp[1]);
-        }
-        return $email;
-    }//end punycodeEmail()
     
-    /**
-     * Creates a nonce that is at least 16 bytes. If a smaller nonce is requested it
-     * will return a 16 byte nonce.
-     *
-     * @param int $bytes The size in bytes of the requested nonce.
-     *
-     * @return string The base64 encoded nonce.
-     */
-    public static function generateNonce(int $bytes = 16): string
-    {
-        if ($bytes < 16) {
-            $bytes = 16;
-        }
-        $raw = random_bytes($bytes);
-        return base64_encode($raw);
-    }//end generateNonce()
     
-    /**
-     * A substitute for the WordPress add_query_arg function.
-     * This function does NOT support user, pass, or fragment.
-     *
-     * @param string      $url              The url to modify.
-     * @param array       $addQueryArgs     Optional. An array of key value pairs.
-     * @param array       $removeQueryArgs  Optional. An array of query args to remove from $url, only key matters.
-     * @param null|string $scheme           Optional. The scheme to use. Only supports http or https.
-     *
-     * @return null|string The url with query args added, or null on failure.
-     */
-    public static function modifyQueryArgs(string $url, array $addQueryArgs = array(), array $removeQueryArgs = array(), $scheme = null)
-    {
-        $parsed = parse_url($url);
-        if (! is_null($scheme)) {
-            $scheme = strtolower($scheme);
-            if (in_array($scheme, array('http', 'https'))) {
-                $parsed['scheme'] = $scheme;
-            }
-        }
-        // todo - throw exception when host is missing
-        
-        if (isset($parsed['query'])) {
-            $queryArray = explode('&', $parsed['query']);
-        } else {
-            $queryArray = array();
-        }
-        $newQueryArray = array();
-        foreach ($queryArray as $string) {
-            $keypair = explode('=', $string);
-            $key = $keypair[0];
-            $value = $keypair[1];
-            if (! in_array($key, $removeQueryArgs)) {
-                $newQueryArray[$key] = $value;
-            }
-        }
-        
-        foreach ($addQueryArgs as $key => $value) {
-            if (! is_bool($value)) {
-                $newQueryArray[$key] = $value;
-            }
-        }
-        if (count($newQueryArray) > 0) {
-            $arr = array();
-            foreach ($newQueryArray as $key => $value) {
-                $arr[] = $key . '=' . $value;
-            }
-            $parsed['query'] = implode('&', $arr);
-        }
-        $url = '';
-        $realurl = '';
-        if (isset($parsed['scheme'])) {
-            $url = $parsed['scheme'] . '://';
-            $realurl = $parsed['scheme'] . '://';
-        }
-        $url = $url . self::punycodeDomain($parsed['host']);
-        $realurl = $realurl . $parsed['host'];
-        if (isset($parsed['port'])) {
-            $url = $url . ':' . $parsed['port'];
-            $realurl = $realurl . ':' . $parsed['port'];
-        }
-        if (! isset($parsed['path'])) {
-            $parsed['path'] = '/';
-        }
-        $url = $url . $parsed['path'];
-        $realurl = $realurl . $parsed['path'];
-        if (count($newQueryArray) > 0) {
-            $url = $url . '?' . $parsed['query'];
-            $realurl = $realurl . '?' . $parsed['query'];
-        }
-        if ($test = filter_var($url, FILTER_VALIDATE_URL)) {
-            return $realurl;
-        }
-        return null;
-    }//end modifyQueryArgs()
-
-    /**
-     * Creates a cryptographically strong 256 bit salt.
-     *
-     * @return string The generated salt, a base64 encoded string.
-     */
-    public static function saltShaker(): string
-    {
-        $raw = random_bytes(32);
-        return base64_encode($raw);
-    }//end saltShaker()
 
     /* Pluggable Methods */
     
     /**
-     * For use with `wp_hash()` pluggable function. Creates a secure hash string if the
+     * For use with `wp_hash()` pluggable function. Creates a secure hash string of the
      * specified number of bytes.
      *
      * If the specified number of bytes is < SODIUM_CRYPTO_GENERICHASH_BYTES_MIN
